decoder -> endecoder, hirarchy change
